requirements.txt
Flask
flask-cors
requests
python-docx
PyPDF2